// src/api/generated/models/post-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 4차 프로젝트 7팀 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuctionRequest } from './auction-request';

/**
 * 
 * @export
 * @interface PostRequest
 */
export interface PostRequest {
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    'content': string;
    /**
     * 
     * @type {number}
     * @memberof PostRequest
     */
    'price': number;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    'category': string;
    /**
     * 
     * @type {string}
     * @memberof PostRequest
     */
    'place': string;
    /**
     * 
     * @type {boolean}
     * @memberof PostRequest
     */
    'saleStatus': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PostRequest
     */
    'auctionStatus': boolean;
    /**
     * 
     * @type {AuctionRequest}
     * @memberof PostRequest
     */
    'auctionRequest'?: AuctionRequest;
}